Changed data type for Asort and added a helper function

+++ Asort.cc
@@ -3,7 +3,7 @@
 
 //Asort function
 void insertionSort(long data[], long n){
-	for(int i = 1, j; i < n; i++){
+	for(long i = 1, j; i < n; i++){
 		long tmp = data[i];
 		for(j = i; j > 0 && tmp < data[j-1]; j--){
 			data[j] = data[j-1];
